# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "在家办公小程序接口文档",
        "version": "1.0"
    },
    "paths": {
...
        "/api/forumcommentlike/save": {
            "post": {
                "tags": [
                    "中心广场"
                ],
                "summary": "增加评论点赞取消点赞",
                "operationId": "App\\Http\\Controllers\\Api\\ForumCommentLikeController::save",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "forum_comment_id": {
                                        "description": "评论id",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "用户id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/forumrecord/save": {
            "post": {
                "tags": [
                    "中心广场"
                ],
                "summary": "增加浏览记录或点赞取消点赞",
                "operationId": "App\\Http\\Controllers\\Api\\ForumRecordController::save",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "forum_id": {
                                        "description": "帖子id",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "用户id",
                                        "type": "integer"
                                    },
                                    "type": {
                                        "description": "类型，1表示浏览，2表示点赞",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/newsrecord/save": {
            "post": {
                "tags": [
                    "新闻"
                ],
                "summary": "增加浏览记录",
                "operationId": "App\\Http\\Controllers\\Api\\NewsRecordController::save",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "news_id": {
                                        "description": "文章id",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "用户id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}